found a way to make the dropdown menu close when you click outside the element
added focus pseudo class to the menu button and set it to display the dropdown-content when it has the class of display on it. the Javascript for the dropdown menu is still required to make it so you can click on the button again to close. I also had to add :focus class for text underline and the: ::after pseudo element