added a media query for desktop to set the :focus styles for only desktop
I did this because on mobile, the :focus pseudo class was causing issues when clicking on a button while another button's dropdown content is active.

What happened was this: I click on one of the buttons and it opens the dropdown content, but then when I click on a different button, it  sometimes would just close the current dropdown menu without opening the next content. by adding a media query for desktop i was able to put all the :focus styles into this media query and then remove the :focus pseudo class for mobile and just have it add the display/active/rotate classes to the menu-button class normally