did a lot of work on the desktop version of the designed-for-the-future section
I gave the image a position of absolute then use transform: translate to move it off to the side and put an overflow: hidden on the main section container.

I also added word spacing to the text to make it look more like the design

I changed the mobile media query to 950px from 850px. this is because this section starts to look bad the smaller it gets so I wanted the breakpoint to happen sooner