started the state-of-the-art section
I had to fuss wit this section a bit to get the image to behave sort of the way I wanted. This project has been hard with the placement of the images. I feel like there is a better way to position them so when the screen is being resized, it wont look like complete garbage.